cheri: Apply the tasklet.cocci script
Run the tasklet.cocci script to fix tasklet calling API.

Coccinelle can do cross-file analysis provided that all the
affected files are in a single invocation. Thus to catch all
matches of the script use the following hack:
- Determine all directories that contain potential matches,
  e.g. by grepping for file that contain the string "tasklet".
- For each directory <DIR> run
	spatch -sp-file scripts/coccinelle/cheri/tasklet.cocci <DIR>/*.[ch]

+++ drivers/crypto/img-hash.c
@@ -355,7 +355,7 @@ static int img_hash_dma_init(struct img_hash_dev *hdev)
 	return 0;
 }
 
-static void img_hash_dma_task(unsigned long d)
+static void img_hash_dma_task(uintptr_t d)
 {
 	struct img_hash_dev *hdev = (struct img_hash_dev *)d;
 	struct img_hash_request_ctx *ctx;
@@ -891,7 +891,7 @@ static int img_unregister_algs(struct img_hash_dev *hdev)
 	return 0;
 }
 
-static void img_hash_done_task(unsigned long data)
+static void img_hash_done_task(uintptr_t data)
 {
 	struct img_hash_dev *hdev = (struct img_hash_dev *)data;
 	int err = 0;
